split up settings into sub files
