+++ scripts/settings/main-settings.ts
@@ -1,10 +1,11 @@
-import { Modal, Plugin, PluginSettingTab, Setting } from 'obsidian';
-import { Utils } from './utils/utils';
-import HTMLExportPlugin from './main';
-import { Path } from './utils/path';
+import { Plugin, PluginSettingTab, Setting } from 'obsidian';
+import { Utils } from '../utils/utils';
+import { Path } from '../utils/path';
 import pluginStylesBlacklist from 'assets/third-party-styles-blacklist.txt';
+import { FlowList } from './flow-list';
 
-export interface ExportSettingsData 
+
+export interface MainSettingsData 
 {
 	// Inlining Options
 	inlineCSS: boolean;
@@ -51,7 +52,7 @@ export interface ExportSettingsData
 	lastExportPath: string;
 }
 
-const DEFAULT_SETTINGS: ExportSettingsData =
+const DEFAULT_SETTINGS: MainSettingsData =
 {
 	// Inlining Options
 	inlineCSS: false,
@@ -98,48 +99,9 @@ const DEFAULT_SETTINGS: ExportSettingsData =
 	lastExportPath: '',
 }
 
-export class FlowList {
-	containerEl: HTMLElement;
-	flowListEl: HTMLElement;
-	checkedList: string[] = [];
+export class MainSettings extends PluginSettingTab {
 
-	constructor(containerEl: HTMLElement) {
-		this.containerEl = containerEl;
-		this.flowListEl = this.containerEl.createDiv({ cls: 'flow-list' });
-
-	}
-
-	addItem(name: string, key: string, value: boolean, onChange: (value: boolean) => void): HTMLElement {
-		let item = this.flowListEl.createDiv({ cls: 'flow-item' });
-		let checkbox = item.createEl('input', { type: 'checkbox' });
-		checkbox.checked = value;
-		if (checkbox.checked) this.checkedList.push(key)
-
-		checkbox.addEventListener('change', (evt) => {
-			if (checkbox.checked) {
-				if (!this.checkedList.includes(key))
-					this.checkedList.push(key)
-			}
-			else {
-				if (this.checkedList.includes(key))
-					this.checkedList.remove(key)
-			}
-		});
-
-		checkbox.addEventListener('change', (evt) => onChange(checkbox.checked));
-
-
-		let label = item.createDiv({ cls: 'flow-label' });
-		label.setText(name);
-
-		return item;
-	}
-
-}
-
-export class ExportSettings extends PluginSettingTab {
-
-	static settings: ExportSettingsData = DEFAULT_SETTINGS;
+	static settings: MainSettingsData = DEFAULT_SETTINGS;
 	static plugin: Plugin;
 
 	private blacklistedPluginIDs: string[] = [];
@@ -152,17 +114,17 @@ export class ExportSettings extends PluginSettingTab {
 
 	constructor(plugin: Plugin) {
 		super(app, plugin);
-		ExportSettings.plugin = plugin;
+		MainSettings.plugin = plugin;
 	}
 
 	static async loadSettings() {
-		ExportSettings.settings = Object.assign({}, DEFAULT_SETTINGS, await ExportSettings.plugin.loadData());
-		ExportSettings.settings.customLineWidth = ExportSettings.settings.customLineWidth.toString();
-		if (ExportSettings.settings.customLineWidth === "0") ExportSettings.settings.customLineWidth = "";
+		MainSettings.settings = Object.assign({}, DEFAULT_SETTINGS, await MainSettings.plugin.loadData());
+		MainSettings.settings.customLineWidth = MainSettings.settings.customLineWidth.toString();
+		if (MainSettings.settings.customLineWidth === "0") MainSettings.settings.customLineWidth = "";
 	}
 
 	static async saveSettings() {
-		await ExportSettings.plugin.saveData(ExportSettings.settings);
+		await MainSettings.plugin.saveData(MainSettings.settings);
 	}
 
 	display() {
@@ -200,20 +162,20 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Include theme toggle')
 			.setDesc('Adds a theme toggle to the left sidebar.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.addDarkModeToggle)
+				.setValue(MainSettings.settings.addDarkModeToggle)
 				.onChange(async (value) => {
-					ExportSettings.settings.addDarkModeToggle = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.addDarkModeToggle = value;
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
 			.setName('Include document outline')
 			.setDesc('Adds the document\'s table of contents to the right sidebar.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.includeOutline)
+				.setValue(MainSettings.settings.includeOutline)
 				.onChange(async (value) => {
-					ExportSettings.settings.includeOutline = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.includeOutline = value;
+					await MainSettings.saveSettings();
 				}
 				));
 
@@ -221,10 +183,10 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Include file tree')
 			.setDesc('Adds an interactive file tree to the left sidebar.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.includeFileTree)
+				.setValue(MainSettings.settings.includeFileTree)
 				.onChange(async (value) => {
-					ExportSettings.settings.includeFileTree = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.includeFileTree = value;
+					await MainSettings.saveSettings();
 				}
 				));
 
@@ -232,10 +194,10 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Add filename as title')
 			.setDesc('If the first header is not an H1, include the file name as a title at the top of the page.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.addFilenameTitle)
+				.setValue(MainSettings.settings.addFilenameTitle)
 				.onChange(async (value) => {
-					ExportSettings.settings.addFilenameTitle = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.addFilenameTitle = value;
+					await MainSettings.saveSettings();
 				}));
 
 		hr = contentEl.createEl("hr");
@@ -252,20 +214,20 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Start Outline Collapsed')
 			.setDesc('Start the document\'s table of contents with all items collapsed')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.startOutlineCollapsed)
+				.setValue(MainSettings.settings.startOutlineCollapsed)
 				.onChange(async (value) => {
-					ExportSettings.settings.startOutlineCollapsed = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.startOutlineCollapsed = value;
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
 			.setName('Allow folding headings')
 			.setDesc('Allow headings to be folded with an arrow icon beside each heading, just as in Obsidian.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.allowFoldingHeadings)
+				.setValue(MainSettings.settings.allowFoldingHeadings)
 				.onChange(async (value) => {
-					ExportSettings.settings.allowFoldingHeadings = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.allowFoldingHeadings = value;
+					await MainSettings.saveSettings();
 				}));
 
 		//#endregion
@@ -286,30 +248,30 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Embed CSS')
 			.setDesc('Embed the CSS into the HTML file.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.inlineCSS)
+				.setValue(MainSettings.settings.inlineCSS)
 				.onChange(async (value) => {
-					ExportSettings.settings.inlineCSS = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.inlineCSS = value;
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
 			.setName('Embed JS')
 			.setDesc('Embed the JS into the HTML file.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.inlineJS)
+				.setValue(MainSettings.settings.inlineJS)
 				.onChange(async (value) => {
-					ExportSettings.settings.inlineJS = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.inlineJS = value;
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
 			.setName('Embed Images')
 			.setDesc('Embed the images into the HTML file.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.inlineImages)
+				.setValue(MainSettings.settings.inlineImages)
 				.onChange(async (value) => {
-					ExportSettings.settings.inlineImages = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.inlineImages = value;
+					await MainSettings.saveSettings();
 				}));
 
 		//#endregion
@@ -329,16 +291,16 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Document Width')
 			.setDesc('Sets the line width of the exported document. Use any css units.\nDefault units: px')
 			.addText((text) => text
-				.setValue(ExportSettings.settings.customLineWidth)
+				.setValue(MainSettings.settings.customLineWidth)
 				.setPlaceholder('Leave blank for default')
 				.onChange(async (value) => {
-					ExportSettings.settings.customLineWidth = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.customLineWidth = value;
+					await MainSettings.saveSettings();
 				}
 				))
 			.addExtraButton((button) => button.setIcon('reset').setTooltip('Reset to default').onClick(() => {
-				ExportSettings.settings.customLineWidth = "";
-				ExportSettings.saveSettings();
+				MainSettings.settings.customLineWidth = "";
+				MainSettings.saveSettings();
 				this.display();
 			}));
 
@@ -346,16 +308,16 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Content Width')
 			.setDesc('Sets the width of the central content section of the document. This will push the sidebars towards the edges of the screen the larger it is leaving margins on either side of the document. Use any css units.\nDefault units: px')
 			.addText((text) => text
-				.setValue(ExportSettings.settings.contentWidth)
+				.setValue(MainSettings.settings.contentWidth)
 				.setPlaceholder('Leave blank for default')
 				.onChange(async (value) => {
-					ExportSettings.settings.contentWidth = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.contentWidth = value;
+					await MainSettings.saveSettings();
 				}
 				))
 			.addExtraButton((button) => button.setIcon('reset').setTooltip('Reset to default').onClick(() => {
-				ExportSettings.settings.contentWidth = "";
-				ExportSettings.saveSettings();
+				MainSettings.settings.contentWidth = "";
+				MainSettings.saveSettings();
 				this.display();
 			}));
 
@@ -363,16 +325,16 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Sidebar Width')
 			.setDesc('Sets the width of the sidebar\'s content. Use any css units.\nDefault units: px')
 			.addText((text) => text
-				.setValue(ExportSettings.settings.sidebarWidth)
+				.setValue(MainSettings.settings.sidebarWidth)
 				.setPlaceholder('Leave blank for default')
 				.onChange(async (value) => {
-					ExportSettings.settings.sidebarWidth = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.sidebarWidth = value;
+					await MainSettings.saveSettings();
 				}
 				))
 			.addExtraButton((button) => button.setIcon('reset').setTooltip('Reset to default').onClick(() => {
-				ExportSettings.settings.sidebarWidth = "";
-				ExportSettings.saveSettings();
+				MainSettings.settings.sidebarWidth = "";
+				MainSettings.saveSettings();
 				this.display();
 			}));
 
@@ -393,13 +355,13 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Dataview Block Render Wait Time')
 			.setDesc('In milliseconds.\n\nWait this long for each dataview block to render. If you have large dataview queries this can help make sure they are rendered correctly.')
 			.addText((text) => text
-				.setValue(ExportSettings.settings.dataviewBlockWaitTime.toString())
+				.setValue(MainSettings.settings.dataviewBlockWaitTime.toString())
 				.setPlaceholder(DEFAULT_SETTINGS.dataviewBlockWaitTime.toString())
 				.onChange(async (value) => {
 					// is the input is not a number then don't let it change
 					if (isNaN(Number(value))) return;
-					ExportSettings.settings.dataviewBlockWaitTime = Number(value);
-					await ExportSettings.saveSettings();
+					MainSettings.settings.dataviewBlockWaitTime = Number(value);
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
@@ -410,31 +372,31 @@ export class ExportSettings extends PluginSettingTab {
 				.addOption('warning', 'Warning')
 				.addOption('error', 'Error')
 				.addOption('fatal', 'Only Fatal Errors')
-				.setValue(ExportSettings.settings.logLevel)
+				.setValue(MainSettings.settings.logLevel)
 				.onChange(async (value: "all" | "warning" | "error" | "fatal" | "none") =>
 				{
-					ExportSettings.settings.logLevel = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.logLevel = value;
+					await MainSettings.saveSettings();
 				}));
 		
 		new Setting(contentEl)
 			.setName('Make names web style')
 			.setDesc('Make the names of files and folders lowercase and replace spaces with dashes.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.makeNamesWebStyle)
+				.setValue(MainSettings.settings.makeNamesWebStyle)
 				.onChange(async (value) => {
-					ExportSettings.settings.makeNamesWebStyle = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.makeNamesWebStyle = value;
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
 			.setName('Beautify HTML')
 			.setDesc('Beautify the HTML text to make it more human readable at the cost of export speed.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.beautifyHTML)
+				.setValue(MainSettings.settings.beautifyHTML)
 				.onChange(async (value) => {
-					ExportSettings.settings.beautifyHTML = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.beautifyHTML = value;
+					await MainSettings.saveSettings();
 				}));
 
 		//#endregion
@@ -467,11 +429,11 @@ export class ExportSettings extends PluginSettingTab {
 			let hasCSS = pluginPath.joinString('styles.css').exists;
 			if (!hasCSS) return;
 
-			let isChecked = ExportSettings.settings.includePluginCSS.match(new RegExp(`^${plugin}`, 'm')) != null;
+			let isChecked = MainSettings.settings.includePluginCSS.match(new RegExp(`^${plugin}`, 'm')) != null;
 
 			pluginsList.addItem(pluginManifest.name, plugin, isChecked, (value) => {
-				ExportSettings.settings.includePluginCSS = pluginsList.checkedList.join('\n');
-				ExportSettings.saveSettings();
+				MainSettings.settings.includePluginCSS = pluginsList.checkedList.join('\n');
+				MainSettings.saveSettings();
 			});
 		});
 
@@ -479,10 +441,10 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Include Svelte CSS')
 			.setDesc('Include the CSS from any plugins that use the svelte framework.')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.includeSvelteCSS)
+				.setValue(MainSettings.settings.includeSvelteCSS)
 				.onChange(async (value) => {
-					ExportSettings.settings.includeSvelteCSS = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.includeSvelteCSS = value;
+					await MainSettings.saveSettings();
 				}));
 
 		//#endregion
@@ -518,10 +480,10 @@ export class ExportSettings extends PluginSettingTab {
 			.setName('Include global graph view')
 			.setDesc('Include an interactive graph view sim of the WHOLE vault similar to obsidian\'s. ')
 			.addToggle((toggle) => toggle
-				.setValue(ExportSettings.settings.includeGraphView)
+				.setValue(MainSettings.settings.includeGraphView)
 				.onChange(async (value) => {
-					ExportSettings.settings.includeGraphView = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.includeGraphView = value;
+					await MainSettings.saveSettings();
 				}));
 
 		new Setting(contentEl)
@@ -534,13 +496,13 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("How much should linked nodes attract each other? This will make the graph appear more clustered.")
 			.addSlider((slider) => slider
 				.setLimits(0, 100, 1)
-				.setValue(ExportSettings.settings.graphAttractionForce / (2 / 100))
+				.setValue(MainSettings.settings.graphAttractionForce / (2 / 100))
 				.setDynamicTooltip()
 				.onChange(async (value) => {
 					// remap to 0 - 2;
 					let remapMultiplier = 2 / 100;
-					ExportSettings.settings.graphAttractionForce = value * remapMultiplier;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphAttractionForce = value * remapMultiplier;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -550,11 +512,11 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("How long should the links between nodes be? The shorter the links the closer connected nodes will cluster together.")
 			.addSlider((slider) => slider
 				.setLimits(0, 100, 1)
-				.setValue(ExportSettings.settings.graphLinkLength)
+				.setValue(MainSettings.settings.graphLinkLength)
 				.setDynamicTooltip()
 				.onChange(async (value) => {
-					ExportSettings.settings.graphLinkLength = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphLinkLength = value;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -564,11 +526,11 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("How much should nodes repel each other? This will make the graph appear more spread out.")
 			.addSlider((slider) => slider
 				.setLimits(0, 100, 1)
-				.setValue(ExportSettings.settings.graphRepulsionForce / 3)
+				.setValue(MainSettings.settings.graphRepulsionForce / 3)
 				.setDynamicTooltip()
 				.onChange(async (value) => {
-					ExportSettings.settings.graphRepulsionForce = value * 3;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphRepulsionForce = value * 3;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -578,13 +540,13 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("How much should nodes be attracted to the center? This will make the graph appear more dense and circular.")
 			.addSlider((slider) => slider
 				.setLimits(0, 100, 1)
-				.setValue(ExportSettings.settings.graphCentralForce / (5 / 100))
+				.setValue(MainSettings.settings.graphCentralForce / (5 / 100))
 				.setDynamicTooltip()
 				.onChange(async (value) => {
 					// remap to 0 - 5;
 					let remapMultiplier = 5 / 100;
-					ExportSettings.settings.graphCentralForce = value * remapMultiplier;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphCentralForce = value * remapMultiplier;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -594,11 +556,11 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("How large should the largest nodes be? Nodes are sized by how many links they have. The larger a node is the more it will attract other nodes. This can be used to create a good grouping around the most important nodes.")
 			.addSlider((slider) => slider
 				.setLimits(3, 15, 1)
-				.setValue(ExportSettings.settings.graphMaxNodeSize)
+				.setValue(MainSettings.settings.graphMaxNodeSize)
 				.setDynamicTooltip()
 				.onChange(async (value) => {
-					ExportSettings.settings.graphMaxNodeSize = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphMaxNodeSize = value;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -608,11 +570,11 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("How small should the smallest nodes be? The smaller a node is the less it will attract other nodes.")
 			.addSlider((slider) => slider
 				.setLimits(3, 15, 1)
-				.setValue(ExportSettings.settings.graphMinNodeSize)
+				.setValue(MainSettings.settings.graphMinNodeSize)
 				.setDynamicTooltip()
 				.onChange(async (value) => {
-					ExportSettings.settings.graphMinNodeSize = value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphMinNodeSize = value;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -622,11 +584,11 @@ export class ExportSettings extends PluginSettingTab {
 			.setDesc("Edges with a length below this threshold will not be rendered, however they will still contribute to the simulation. This can help large tangled graphs look more organised. Hovering over a node will still display these links.")
 			.addSlider((slider) => slider
 				.setLimits(0, 100, 1)
-				.setValue(100 - ExportSettings.settings.graphEdgePruning)
+				.setValue(100 - MainSettings.settings.graphEdgePruning)
 				.setDynamicTooltip()
 				.onChange(async (value) => {
-					ExportSettings.settings.graphEdgePruning = 100 - value;
-					await ExportSettings.saveSettings();
+					MainSettings.settings.graphEdgePruning = 100 - value;
+					await MainSettings.saveSettings();
 				})
 				.showTooltip()
 			);
@@ -636,175 +598,5 @@ export class ExportSettings extends PluginSettingTab {
 
 		//#endregion
 
-	}
-}
-
-export class ExportModal extends Modal {
-	static isClosed: boolean = true;
-	static canceled: boolean = true;
-	static filePickerModal: HTMLElement;
-
-	constructor() {
-		super(app);
-	}
-
-	/**
-	 * @brief Opens the modal and async blocks until the modal is closed.
-	 * @returns True if the EXPORT button was pressed, false is the export was canceled.
-	 * @override
-	*/
-	async open(): Promise<{ canceled: boolean }> {
-		ExportModal.isClosed = false;
-		ExportModal.canceled = true;
-
-		super.open();
-
-		const { contentEl } = this;
-
-		contentEl.empty();
-
-		this.titleEl.setText('Export to HTML');
-
-		if (HTMLExportPlugin.updateInfo.updateAvailable) {
-			// create red notice showing the update is available
-			let updateNotice = contentEl.createEl('strong', { text: `Update Available: ${HTMLExportPlugin.updateInfo.currentVersion} ⟶ ${HTMLExportPlugin.updateInfo.latestVersion}` });
-			updateNotice.setAttribute("style",
-				`margin-block-start: calc(var(--h3-size)/2);
-			background-color: var(--interactive-normal);
-			padding: 4px;
-			padding-left: 1em;
-			padding-right: 1em;
-			color: var(--color-red);
-			border-radius: 5px;
-			display: block;
-			width: fit-content;`)
-
-			// create normal block with update notes
-			let updateNotes = contentEl.createEl('div', { text: HTMLExportPlugin.updateInfo.updateNote });
-			updateNotes.setAttribute("style",
-				`margin-block-start: calc(var(--h3-size)/2);
-			background-color: var(--background-secondary-alt);
-			padding: 4px;
-			padding-left: 1em;
-			padding-right: 1em;
-			color: var(--text-normal);
-			font-size: var(--font-ui-smaller);
-			border-radius: 5px;
-			display: block;
-			width: fit-content;
-			white-space: pre-wrap;`)
-		}
-
-		let hr = contentEl.createEl("hr");
-		hr.style.marginTop = "20px";
-		hr.style.marginBottom = "20px";
-		hr.style.borderColor = "var(--color-accent)";
-		hr.style.opacity = "0.5";
-
-		contentEl.createEl('h3', { text: 'Basic Options:' });
-
-		new Setting(contentEl)
-			.setName('Export Presets')
-			.setHeading()
-			.addDropdown((dropdown) => dropdown
-				.addOption('website', 'Multi-File Website')
-				.addOption('documents', 'Self-contained Documents')
-				.setValue(ExportSettings.settings.exportPreset)
-				.onChange(async (value) => 
-				{
-					ExportSettings.settings.exportPreset = value;
-
-					switch (value) {
-						case 'documents':
-							ExportSettings.settings.inlineCSS = true;
-							ExportSettings.settings.inlineJS = true;
-							ExportSettings.settings.inlineImages = true;
-							ExportSettings.settings.makeNamesWebStyle = false;
-							ExportSettings.settings.includeGraphView = false;
-							ExportSettings.settings.includeFileTree = false;
-							await ExportSettings.saveSettings();
-
-							break;
-						case 'website':
-							ExportSettings.settings.inlineCSS = false;
-							ExportSettings.settings.inlineJS = false;
-							ExportSettings.settings.inlineImages = false;
-							ExportSettings.settings.makeNamesWebStyle = true;
-							ExportSettings.settings.includeGraphView = false;
-							ExportSettings.settings.includeFileTree = true;
-							await ExportSettings.saveSettings();
-
-							break;
-					}
-
-					this.open();
-				}
-				));
-
-		contentEl.createDiv().outerHTML = 
-		`
-		<div class="setting-item-description" style="white-space: pre-wrap; margin-bottom: 1em;
-		">Multi-File Website: For multiple files as a website.
-Self-contained Documents: For documents which should each be self contained as one file.
-
-<em>For more control open the plugin settings from the button at the bottom of this popup.</em></div>`
-
-		new Setting(contentEl)
-			.setName('Open after export')
-			.addToggle((toggle) => toggle
-				.setTooltip('Open the exported file after exporting.')
-				.setValue(ExportSettings.settings.openAfterExport)
-				.onChange(async (value) => {
-					ExportSettings.settings.openAfterExport = value;
-					await ExportSettings.saveSettings();
-				}));
-
-		new Setting(contentEl)
-			.setName('Only Export Modified')
-			.setDesc('Disable this to do a full re-export.')
-			.addToggle((toggle) => toggle
-			.setValue(ExportSettings.settings.incrementalExport)
-			.onChange(async (value) => {
-				ExportSettings.settings.incrementalExport = value;
-				await ExportSettings.saveSettings();
-			}));
-
-		new Setting(contentEl)
-			.setName('')
-			.setHeading()
-			.addButton((button) => 
-			{
-				button.setButtonText('Export').onClick(async () => 
-				{
-					ExportModal.canceled = false;
-					this.close();
-				});
-
-				button.buttonEl.style.marginRight = 'auto';
-				button.buttonEl.style.marginLeft = 'auto';
-				button.buttonEl.style.width = '-webkit-fill-available';
-				button.buttonEl.style.marginBottom = '2em';
-			});
-
-		new Setting(contentEl)
-			.setDesc("More options located on the plugin settings page.")
-			.addExtraButton((button) => button.setTooltip('Open plugin settings').onClick(() => {
-				//@ts-ignore
-				app.setting.open();
-				//@ts-ignore
-				app.setting.openTabById('webpage-html-export');
-			}));
-
-
-		await Utils.waitUntil(() => ExportModal.isClosed, 60 * 60 * 1000, 10);
-
-		return { canceled: ExportModal.canceled };
-	}
-
-	onClose() {
-		const { contentEl } = this;
-		contentEl.empty();
-		ExportModal.isClosed = true;
-		ExportModal.filePickerModal.remove();
 	}
 }
